Move logic to OperationPreferredDateUpdaterUseCase
